refactor(migration): extract migration modal
