+++ packages/bruno-app/src/components/RequestTabs/CollectionHeader/index.js
@@ -23,7 +23,8 @@ import OpenAPISyncIcon from 'components/Icons/OpenAPISync';
 import { switchWorkspace, renameWorkspaceAction, exportWorkspaceAction, confirmWorkspaceCreation, cancelWorkspaceCreation } from 'providers/ReduxStore/slices/workspaces/actions';
 import { updateWorkspace } from 'providers/ReduxStore/slices/workspaces';
 import { showInFolder } from 'providers/ReduxStore/slices/collections/actions';
-import { toggleCollectionFileMode, toggleAppMode, updateSettingsSelectedTab } from 'providers/ReduxStore/slices/collections';
+import { toggleCollectionFileMode, toggleAppMode } from 'providers/ReduxStore/slices/collections';
+import MigrateToYmlModal from 'components/CollectionSettings/Overview/Migration/MigrateToYmlModal';
 import { findItemInCollection, findItemInCollectionByPathname } from 'utils/collections';
 import find from 'lodash/find';
 import get from 'lodash/get';
@@ -270,16 +271,7 @@ const CollectionHeader = ({ collection, isScratchCollection }) => {
     );
   };
 
-  const viewMigrationSettings = () => {
-    dispatch(
-      addTab({
-        uid: collection.uid,
-        collectionUid: collection.uid,
-        type: 'collection-settings'
-      })
-    );
-    dispatch(updateSettingsSelectedTab({ collectionUid: collection.uid, tab: 'overview' }));
-  };
+  const [showMigrateModal, setShowMigrateModal] = useState(false);
 
   const viewOpenApiSync = () => {
     dispatch(addTab({
@@ -708,7 +700,7 @@ const CollectionHeader = ({ collection, isScratchCollection }) => {
                 <button
                   type="button"
                   className="pill-main"
-                  onClick={viewMigrationSettings}
+                  onClick={() => setShowMigrateModal(true)}
                 >
                   <IconTransform size={13} strokeWidth={1.5} />
                   <span className="pill-label">Migrate to YML</span>
@@ -760,6 +752,12 @@ const CollectionHeader = ({ collection, isScratchCollection }) => {
           </div>
         )}
       </div>
+      {showMigrateModal && (
+        <MigrateToYmlModal
+          collection={collection}
+          onClose={() => setShowMigrateModal(false)}
+        />
+      )}
     </StyledWrapper>
   );
 };
